improved the ui

+++ server/controllers/userController.js
@@ -1,23 +1,24 @@
-const User = require('../models/userModel');
-const Post = require('../models/postModel');
-const { FILE_MAX_SIZE } = require('../utils/constants');
-const fs = require('fs').promises;
-const sharp = require('sharp');
-const multer = require('multer');
-const jwt = require('jsonwebtoken');
-const renderRes = require('../utils/renderRes');
-const { isEmailAlreadyTaken, isNameAlreadyTaken } = require('../utils/utils');
+const User = require('../models/userModel')
+const Comment = require('../models/commentModel')
+const Post = require('../models/postModel')
+const { FILE_MAX_SIZE } = require('../utils/constants')
+const fs = require('fs').promises
+const sharp = require('sharp')
+const multer = require('multer')
+const jwt = require('jsonwebtoken')
+const renderRes = require('../utils/renderRes')
+const { isEmailAlreadyTaken, isNameAlreadyTaken } = require('../utils/utils')
 
-const multerStorage = multer.memoryStorage();
+const multerStorage = multer.memoryStorage()
 const multerFilter = (req, file, cb) => {
-  console.log('File', file);
-  if (file.mimetype.startsWith('image')) cb(null, true);
-  else cb(new Error('Not an image! Please upload only images.'), false);
-};
+  console.log('File', file)
+  if (file.mimetype.startsWith('image')) cb(null, true)
+  else cb(new Error('Not an image! Please upload only images.'), false)
+}
 const upload = multer({
   fileFilter: multerFilter,
   storage: multerStorage,
-});
+})
 
 exports.upload = (req, res, next) => {
   upload.single('photo')(req, res, async (err) => {
@@ -29,75 +30,74 @@ exports.upload = (req, res, next) => {
         res,
         status: 400,
         message: 'Image size should be less than 5 MB',
-      });
-    if (err) return renderRes({ res, status: 400, message: err.message });
+      })
+    if (err) return renderRes({ res, status: 400, message: err.message })
     req.photo = req.file
       ? `photo-${req.currentUser._id}-${Date.now()}.${
           req.file.mimetype.split('/')[1]
         }`
-      : undefined;
-    next();
-  });
-};
+      : undefined
+    next()
+  })
+}
 
 exports.getCurrentUser = async (req, res) => {
   // 1) Getting toke
-  let token;
+  let token
   if (req.headers.authorization?.startsWith('Bearer'))
-    token = req.headers.authorization.split(' ')[1];
-  if (req.cookies?.jwt) token = req.cookies.jwt;
-  console.log(req.cookies);
+    token = req.headers.authorization.split(' ')[1]
+  if (req.cookies?.jwt) token = req.cookies.jwt
+  console.log(req.cookies)
   try {
     // 2) Cheking if the token exists
     if (!token)
-      throw new Error('You are not logged in! Please log in to get access.');
+      throw new Error('You are not logged in! Please log in to get access.')
     // 3) Verifying the token
-    const decoded = jwt.verify(token, process.env.JWT_SECRET);
+    const decoded = jwt.verify(token, process.env.JWT_SECRET)
     // 4) Check if user still exists
-    console.log(decoded);
-    const user = await User.findById(decoded.id);
+    console.log(decoded)
+    const user = await User.findById(decoded.id)
     if (!user)
-      throw new Error('The user belonging to this token does no longer exist');
+      throw new Error('The user belonging to this token does no longer exist')
     // if it exists
-    renderRes({ res, status: 200, data: { user } });
+    renderRes({ res, status: 200, data: { user } })
   } catch (err) {
-    renderRes({ res, status: 401, message: err.message });
+    renderRes({ res, status: 401, message: err.message })
   }
-};
+}
 exports.updateMe = async (req, res) => {
-  const { email, fullName } = req.body;
-  const updateObj = { email, fullName };
-  if (req.photo) updateObj.photo = req.photo;
+  const { email, fullName } = req.body
+  const updateObj = { email, fullName }
+  if (req.photo) updateObj.photo = req.photo
   try {
     const user = await User.findByIdAndUpdate(req.currentUser._id, updateObj, {
       runValidators: true,
-    });
+    })
     // upload image
     if (req.photo) {
-      const path = 'public/img/users';
+      const path = 'public/img/users'
       await sharp(req.file.buffer)
         .resize(500, 500)
-        .toFile(`${path}/${req.photo}`);
-      if (user.photo !== 'default.jpg')
-        await fs.unlink(`${path}/${user.photo}`);
+        .toFile(`${path}/${req.photo}`)
+      if (user.photo !== 'default.jpg') await fs.unlink(`${path}/${user.photo}`)
     }
     //
-    renderRes({ res, status: 201, data: { user } });
+    renderRes({ res, status: 201, data: { user } })
   } catch (err) {
-    const error = isNameAlreadyTaken(err) || isEmailAlreadyTaken(err) || err;
+    const error = isNameAlreadyTaken(err) || isEmailAlreadyTaken(err) || err
     renderRes({
       res,
       status: 400,
       message: error.message,
       errors: err.errors,
-    });
+    })
   }
-};
+}
 exports.updatePassword = async (req, res) => {
-  const { currentPassword, newPassword, confirmPassword } = req.body;
-  console.log({ currentPassword, newPassword, confirmPassword });
+  const { currentPassword, newPassword, confirmPassword } = req.body
+  console.log({ currentPassword, newPassword, confirmPassword })
   try {
-    const user = await User.findById(req.currentUser._id).select('+password');
+    const user = await User.findById(req.currentUser._id).select('+password')
     if (!(await user.correctPassword(currentPassword || '', user.password)))
       return renderRes({
         res,
@@ -105,21 +105,21 @@ exports.updatePassword = async (req, res) => {
         message: currentPassword
           ? 'Current password is wrong'
           : 'Current password is required',
-      });
+      })
     // if so
-    user.password = newPassword;
-    user.passwordConfirm = confirmPassword;
-    await user.save();
-    renderRes({ res, status: 200, data: { user: req.currentUser } });
+    user.password = newPassword
+    user.passwordConfirm = confirmPassword
+    await user.save()
+    renderRes({ res, status: 200, data: { user: req.currentUser } })
   } catch (err) {
-    renderRes({ res, status: 400, message: err.message, errors: err.errors });
+    renderRes({ res, status: 400, message: err.message, errors: err.errors })
   }
-};
+}
 exports.deleteMe = async (req, res) => {
-  const { currentPassword } = req.body;
+  const { currentPassword } = req.body
   try {
     // check if currentPassword is correct
-    const user = await User.findById(req.currentUser._id).select('+password');
+    const user = await User.findById(req.currentUser._id).select('+password')
     if (!(await user.correctPassword(currentPassword || '', user.password)))
       return renderRes({
         res,
@@ -127,20 +127,22 @@ exports.deleteMe = async (req, res) => {
         message: currentPassword
           ? 'Current password is wrong'
           : 'Current password is required',
-      });
+      })
     // if so
     // 2- delete all user posts
-    await Post.deleteMany({ author: req.currentUser._id });
+    await Post.deleteMany({ author: req.currentUser._id })
+    // delete all user comments
+    await Comment.deleteMany({ user: req.currentUser._id })
     // 2- delete user
-    await User.findByIdAndDelete(req.currentUser._id);
+    await User.findByIdAndDelete(req.currentUser._id)
     // 3- clear cookie
-    res.clearCookie('jwt');
+    res.clearCookie('jwt')
     renderRes({
       res,
       status: 200,
       data: { message: 'User account successfully deleted' },
-    });
+    })
   } catch (err) {
-    renderRes({ res, status: 400, message: err.message, errors: err.errors });
+    renderRes({ res, status: 400, message: err.message, errors: err.errors })
   }
-};
+}
